requirements.txt
fastapi==0.115.2
motor==3.6.0
uvicorn==0.31.1
pydantic-settings==2.5.2
